Refactor GridMenu
