+++ packages/grid/_modules_/grid/components/menu/GridMenu.tsx
@@ -42,37 +42,15 @@ export const GridMenu: React.FC<MenuProps> = ({
   position,
   ...other
 }) => {
-  const menuListRef = React.useRef<HTMLUListElement | null>(null);
-  const focusTimeout = React.useRef<any>();
-
-  React.useEffect(() => {
-    focusTimeout.current = setTimeout(() => {
-      if (open && menuListRef.current && target) {
-        const firstItem = menuListRef.current.querySelector('li:first-child');
-        if (firstItem) {
-          (firstItem as HTMLLIElement).focus();
-        } else {
-          menuListRef.current.focus();
-        }
-      }
-    }, 10);
-  }, [target, open, menuListRef]);
-
-  React.useEffect(() => {
-    return () => {
-      clearTimeout(focusTimeout.current);
-    };
-  }, []);
-
   return (
     <Popper open={open} anchorEl={target as any} transition placement={position} {...other}>
       {({ TransitionProps, placement }) => (
         <Grow {...TransitionProps} style={{ transformOrigin: transformOrigin[placement] }}>
           <Paper>
             <ClickAwayListener onClickAway={onClickAway}>
-              <MenuList ref={menuListRef} id="menu-list-grow" onKeyDown={onKeyDown}>
+              <div>
                 {children}
-              </MenuList>
+              </div>
             </ClickAwayListener>
           </Paper>
         </Grow>
